% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp.R
\name{ctpp_tblsearch}
\alias{ctpp_tblsearch}
\title{Search CTPP table codes}
\usage{
ctpp_tblsearch(prefix, regex, year)
}
\arguments{
\item{prefix}{2-character pattern: see \code{vignette("psrcctpp_basics", package = "psrcctpp")}}

\item{regex}{string pattern to match table description}

\item{year}{last of 5-year CTPP span, e.g. 2016 for ctpp1216 survey}
}
\value{
data table
}
\description{
Search CTPP table codes
}
