% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp.R
\name{ctpp_varsearch}
\alias{ctpp_varsearch}
\title{Search CTPP variable codes}
\usage{
ctpp_varsearch(table_code, year)
}
\arguments{
\item{table_code}{requested data table as string, e.g. "A302103"}

\item{year}{last of 5-year CTPP span, e.g. 2016 for ctpp1216 survey}
}
\value{
data table
}
\description{
Search CTPP variable codes
}
