% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp.R
\name{fetch_ctpp_from_api}
\alias{fetch_ctpp_from_api}
\title{Fetch CTPP data from AASHTO api
Helper to get_psrc_ctpp}
\usage{
fetch_ctpp_from_api(scale, table_code, dyear, geoids)
}
\arguments{
\item{scale}{"county", "place", or "tract" for residence/workplace tables;
"county-county", "place-place", "place-county", "county-place", or "tract-tract" for O-D tables}

\item{table_code}{requested data table as string, e.g. "A302103"}

\item{dyear}{last of 5-year CTPP span, e.g. 2016 for ctpp1216 survey}

\item{geoids}{optional string vector of GEOID codes to limit the table}
}
\value{
data table
}
\description{
Fetch CTPP data from AASHTO api
Helper to get_psrc_ctpp
}
