% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp.R
\name{fetch_ctpp_from_file}
\alias{fetch_ctpp_from_file}
\title{Fetch ftp-format CTPP data from network location
Helper to get_psrc_ctpp}
\usage{
fetch_ctpp_from_file(scale, table_code, dyear = 2016, filepath = "default")
}
\arguments{
\item{scale}{"county", "place", or "tract" for residence/workplace tables;
"county-county", "place-place", "place-county", "county-place", or "tract-tract" for O-D tables}

\item{table_code}{requested data table as string, e.g. "A302103"}

\item{dyear}{last of 5-year CTPP span, e.g. 2016 for ctpp1216 survey}

\item{filepath}{optional network location of downloaded ftp CTPP files}
}
\value{
data table
}
\description{
Fetch ftp-format CTPP data from network location
Helper to get_psrc_ctpp
}
