% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp.R
\name{scale_code_lookup}
\alias{scale_code_lookup}
\title{Correspondence table for CTPP codes, table type, and scale}
\usage{
scale_code_lookup()
}
\value{
correspondence table
}
\description{
Correspondence table for CTPP codes, table type, and scale
}
