% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp.R
\name{api_gofer}
\alias{api_gofer}
\title{Fetch API results
Helper function sending the API call & parsing its response}
\usage{
api_gofer(url)
}
\arguments{
\item{url}{api call}
}
\value{
data table
}
\description{
Fetch API results
Helper function sending the API call & parsing its response
}
