% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp.R
\name{ctpp_shares}
\alias{ctpp_shares}
\title{Add shares to CTPP table}
\usage{
ctpp_shares(df)
}
\arguments{
\item{df}{dataframe, either result of get_psrc_ctpp or psrc_ctpp_stat}
}
\value{
dataframe with share & share_moe
}
\description{
Add shares to CTPP table
}
