% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp.R
\name{ctpp_stat}
\alias{ctpp_stat}
\alias{psrc_ctpp_sum}
\title{CTPP summary statistics}
\usage{
psrc_ctpp_sum(ctpp_data, group_vars = NULL, incl_na = TRUE)
}
\arguments{

\item{df}{result of get_psrc_ctpp function}

\item{group_vars}{grouping variables, original or added}

\item{incl_na}{whether to include NA lines}
}
\value{
A table with the variable names and labels, summary statistic and margin of error
}
\description{
CTPP summary statistics

Generate CTPP sums
}
