% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp.R
\name{get_psrc_ctpp}
\alias{get_psrc_ctpp}
\title{Retrieve CTPP data}
\usage{
get_psrc_ctpp(dyear = 2016, data_table, scale, geoids = NULL)
}
\arguments{
\item{dyear}{last of 5-year CTPP span, e.g. 2016 for ctpp1216 survey}

\item{data_table}{requested data table as string, e.g. "A302103"}

\item{scale}{"state", "county", "place", or "tract" for residence/workplace tables; "state-state", "county-county", "place-place", "place-county", "county-place", or "tract-tract" for O-D tables}

\item{geoids}{optional string vector of GEOID codes to limit the table}
}
\value{
data table
}
\description{
Retrieve CTPP data
}
