% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp_api.R
\name{get_psrc_ctpp_api}
\alias{get_psrc_ctpp_api}
\title{Retrieve CTPP data}
\usage{
get_psrc_ctpp_api(table_code, dyear = 2016, scale, geoids = NULL)
}
\arguments{
\item{table_code}{either a table code as string--e.g. "A302103"--or string vector of variable codes}

\item{dyear}{last of 5-year CTPP span, e.g. 2016 for ctpp1216 survey}

\item{scale}{"county", "place", or "tract" for residence/workplace tables; "county-county", "place-place", "place-county", "county-place", or "tract-tract" for O-D tables}

\item{geoids}{optional string vector of GEOID codes to limit the table}
}
\value{
data table
}
\description{
Retrieve CTPP data
}
