% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp.R
\name{psrc_ctpp_stat}
\alias{psrc_ctpp_stat}
\title{Generic call for CTPP summary statistics}
\usage{
psrc_ctpp_stat(df, group_vars, stat_type = "sum", incl_na = FALSE)
}
\arguments{
\item{df}{result of get_psrc_ctpp function}

\item{group_vars}{grouping variables, original or added}

\item{stat_type}{for now, "sum" is only option}

\item{incl_na}{whether to include NA lines}
}
\description{
Given specific form by related \code{\link{ctpp_stat}} functions.
}
