% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_ctpp.R
\name{scale_code_lookup}
\alias{scale_code_lookup}
\title{Correspondence table for CTPP codes, table type, and scale}
\usage{
scale_code_lookup(scale, table_code)
}
\arguments{
\item{scale}{"county", "place", or "tract" for residence/workplace tables;
"county-county", "place-place", "place-county", "county-place", or "tract-tract" for flow tables}

\item{table_code}{requested data table as string, e.g. "A302103"}
}
\value{
correspondence table
}
\description{
Correspondence table for CTPP codes, table type, and scale
}
